# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Notification API",
        "description": "API for managing notifications",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Get list of notifications",
                "description": "Returns list of notifications",
                "operationId": "getNotificationsList",
                "parameters": [
                    {
                        "name": "recipient",
                        "in": "query",
                        "description": "Filter by recipient",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "service_name",
                        "in": "query",
                        "description": "Filter by service name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Store new notification",
                "description": "Returns notification data",
                "operationId": "storeNotification",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "channel",
                                    "recipient",
                                    "message"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "email"
                                    },
                                    "channel": {
                                        "type": "string",
                                        "example": "email"
                                    },
                                    "recipient": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "subject": {
                                        "type": "string",
                                        "example": "Hello"
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "World"
                                    },
                                    "service_name": {
                                        "type": "string",
                                        "example": "MyService"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/v1/notifications/{id}": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Get notification information",
                "description": "Returns notification data",
                "operationId": "getNotificationById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Notification id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/notifications/{id}/read": {
            "patch": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Mark notification as read",
                "description": "Returns updated notification",
                "operationId": "markNotificationAsRead",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Notification id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Notifications",
            "description": "Notifications"
        }
    ]
}